refactor: implement createVault functionality in WriteStorageProvider and remove CreateVaultProvider to streamline vault creation process across clients
