+++ task.md
@@ -1,3 +1,6 @@
 # Notes for the task related to the current pull request.
 
 The content of this file should be cleared on Pull Request submit.
+
+- [ ] createVaultDefaultCoins as part of the createVault mutation
+- [ ] Shared currentVaultId corrector
